cleaned up code and made small parameter changes
