+++ navigation/search.py
@@ -4,11 +4,14 @@ from navigation.coordinate_utils import d_calc
 from mrover.msg import GPSPointList, WaypointType
 from state_machine.state import State
 from . import recovery, waypoint
+from navigation.coordinate_utils import gen_marker
 from .context import convert_cartesian_to_gps, Context
 from .trajectory import SearchTrajectory
 from navigation import approach_target
 from rclpy.time import Time
 from rclpy.duration import Duration
+from visualization_msgs.msg import Marker
+from rclpy.publisher import Publisher
 
 
 class SearchState(State):
@@ -19,10 +22,12 @@ class SearchState(State):
     time_begin: Optional[Time]
     time_last_updated: Time
     total_distance: float
+    marker_pub: Publisher
 
     def on_enter(self, context: Context) -> None:
         if SearchState.trajectory is None:
             self.new_trajectory(context)
+        self.marker_pub = context.node.create_publisher(Marker, "spiral_points", 10)
         self.prev_pos = None
         self.time_begin = None
         self.time_last_updated = context.node.get_clock().now()
@@ -49,7 +54,11 @@ class SearchState(State):
             total_time = (context.node.get_clock().now() - self.time_begin).nanoseconds / 1e9
             context.node.get_logger().info(f"Total Distance Traveled: {self.total_distance}m\nTotal Time: {total_time}s\nAverage Speed: {self.total_distance/total_time}m/s")
             self.time_last_updated = context.node.get_clock().now()
-
+            start_pt = self.trajectory.cur_pt - 6 if self.trajectory.cur_pt - 3 >= 0 else self.trajectory.cur_pt
+            end_pt = self.trajectory.cur_pt + 6 if self.trajectory.cur_pt + 3 < len(self.trajectory.coordinates) else len(self.trajectory.coordinates)
+            for i, coord in enumerate(self.trajectory.coordinates[start_pt:end_pt]):
+                self.marker_pub.publish(gen_marker(context=context, point=coord, color=[1.0,0.0,0.0], id=i))
+        
         # Continue executing the path from wherever it left off
         target_position_in_map = SearchState.trajectory.get_current_point()
         cmd_vel, arrived = context.drive.get_drive_command(
