+++ perception/cost_map/cost_map.processing.cpp
@@ -78,7 +78,7 @@ namespace mrover {
 
             
             constexpr double MAX_COST_RADIUS = 2; 
-            constexpr double INFLATION_RADIUS = 4; 
+            constexpr double INFLATION_RADIUS = 3; 
 
             nav_msgs::msg::OccupancyGrid postProcessed = mGlobalGridMsg;
             std::ptrdiff_t width = postProcessed.info.width;
