+++ scripts/debug_course_publisher.py
@@ -63,8 +63,8 @@ class DebugCoursePublisher(Node):
                 convert_waypoint_to_gps(ref_point, waypoint)
                 for waypoint in [
                     (
-                        Waypoint(type=WaypointType(val=WaypointType.WATER_BOTTLE)),
-                        SE3.from_position_orientation(0.0, 0.0),
+                        Waypoint(type=WaypointType(val=WaypointType.NO_SEARCH)),
+                        SE3.from_position_orientation(10.0, 0.0),
                     ),
                     # (
                     #     Waypoint(type=WaypointType(val=WaypointType.NO_SEARCH)),
